add chatview
