+++ src/AntSK/Pages/ChatPage/Components/ChatView.razor
@@ -5,8 +5,11 @@
 @layout OpenLayout
 @inherits AntDomComponentBase
 
-<div id="chat" style="display:flex; flex-direction:column; height:100%; overflow-x:hidden;">
-    <PageHeader Class="site-page-header" Title="@app.Name" Subtitle="@app.Describe" />
+<div id="chat">
+    @if (ShowTitle)
+    {
+        <PageHeader Class="site-page-header" Title="@app.Name" Subtitle="@app.Describe" />
+    }
     <div id="scrollDiv" style="flex:1; width:100%; overflow-y:auto; overflow-x:hidden;padding:10px;">
         <Virtualize Items="@(MessageList.OrderBy(o => o.CreateTime).ToList())" Context="item">
             @if (item.IsSend)
@@ -15,7 +18,17 @@
                     <GridCol Span="23">
                         <div class="chat-bubble sent">
                             <Popover Title="@item.CreateTime.ToString()">
-                                @(item.Context)
+                                <Unbound>
+                                    <Flex Vertical RefBack="context">
+                                        @if (item.FileName != null)
+                                        {
+                                            <p class="message-file">
+                                                <Upload DefaultFileList="[new(){ FileName= item.FileName }]" />
+                                            </p>
+                                        }
+                                        <p>@(item.Context)</p>
+                                    </Flex>
+                                </Unbound>
                             </Popover>
                         </div>
                         <Icon Style="float:right;margin-top:10px;" Type="copy" Theme="outline" OnClick="async () =>await OnCopyAsync(item)" />
@@ -105,6 +118,10 @@
         background-color: #daf8cb;
         align-self: flex-end;
         float: right;
+    }
+
+    .ant-card-body {
+        height: 90% !important;
     }
 </style>
