--- src/AntSK/Pages/ChatPage/Components/ChatView.razor.css
@@ -1,3 +1,0 @@
-.ant-card-body {
-    height: 90% !important;
-}
