+++ src/AntSK/Pages/ChatPage/OpenChat.razor
@@ -6,10 +6,20 @@
 @page "/OpenChat/{AppId}"
 @layout OpenLayout
 
-<ChatView AppId="@AppId">
+<ChatView AppId="@AppId" ShowTitle=true>
 
 </ChatView>
 
+<style>
+    #chat{
+        height:100% ;
+        display:flex;
+        flex-direction: column;
+        overflow-x: hidden;
+        font-family: Arial, sans-serif;
+        margin: 0;
+    }
+</style>
 @code{
     [Parameter]
     public string AppId { get; set; }
